requirements.txt
fastapi
uvicorn
python-multipart
torch
opencv-python
pandas
requests
Pillow
yolov5